+++ pyswarms/multi/__init__.py
@@ -1,1 +1,0 @@
-""" Multi-objective optimization package  for `pyswarms` """