+++ pyswarms/utils/search/base_search.py
@@ -27,7 +27,7 @@ class SearchBase(object):
 
     def __init__(self, optimizer, n_particles, dimensions, options,
                  objective_func, iters,
-                 bounds=None, velocity_clamp=None):
+                 bounds=None, velocity_clamp=(0,1)):
         """Initializes the Search.
 
         Attributes